ref: Code tarmed non checké
git-svn-id: svn://svn.code.sf.net/p/mediboard/code/trunk@14343 a21d6be2-e512-0410-a060-91f5a2946631

+++ modules/dPccam/classes/CCodable.class.php
@@ -82,14 +82,16 @@ class CCodable extends CMbObject {
     }
     $this->_tokens_ngap = "";
     
-    // Suppression des anciens actes Tarmed
-    $this->loadRefsActesTarmed();
-    foreach ($this->_ref_actes_tarmed as $acte) { 
-      if ($msg = $acte->delete()) {
-        return $msg;
+    if(CModule::getInstalled("tarmed")){
+      // Suppression des anciens actes Tarmed
+      $this->loadRefsActesTarmed();
+      foreach ($this->_ref_actes_tarmed as $acte) { 
+        if ($msg = $acte->delete()) {
+          return $msg;
+        }
       }
+      $this->_tokens_tarmed = "";
     }
-    $this->_tokens_tarmed = "";
   }  
   
   /**
